+++ andres_velasquez/dags/sql/query_jujuy.sql
