+++ andres_velasquez/dags/sql/query_palermo.sql
